update portfolio history

+++ Vakapay.Models/Repositories/IPortfolioHistoryRepository.cs
@@ -7,7 +7,7 @@ namespace Vakapay.Models.Repositories
 {
     public interface IPortfolioHistoryRepository : IRepositoryBase<PortfolioHistory>
     {
-        List<PortfolioHistory> FindByUserId(string userId, int from, int to);
+        List<PortfolioHistory> FindByUserId(string userId, long from, long to);
         ReturnObject InsertWithPrice(string userId, string vkcPrice, string btcPrice, string ethPrice, string eosPrice);
     }
 }
